refactor: extract notion modules (templates, flowSteps, instances)
